added running scan loader to drawer btn
