+++ frontend/src/App.vue
@@ -1,93 +1,17 @@
 <script setup lang="ts">
 import Notification from "@/components/common/Notification.vue";
 import api from "@/services/api/index";
-import socket from "@/services/socket";
 import storeConfig from "@/stores/config";
-import storeGalleryFilter from "@/stores/galleryFilter";
 import storeHeartbeat from "@/stores/heartbeat";
 import storeNotifications from "@/stores/notifications";
-import storeRoms, { type SimpleRom } from "@/stores/roms";
-import storeScanning from "@/stores/scanning";
-import type { Events } from "@/types/emitter";
-import { normalizeString } from "@/utils";
-import type { Emitter } from "mitt";
 import { storeToRefs } from "pinia";
-import { inject, onBeforeUnmount, onMounted } from "vue";
+import { onMounted } from "vue";
 
 // Props
-const scanningStore = storeScanning();
 const notificationStore = storeNotifications();
 const { notifications } = storeToRefs(notificationStore);
-const { scanningPlatforms } = storeToRefs(scanningStore);
-const romsStore = storeRoms();
-const galleryFilter = storeGalleryFilter();
-const isFiltered = normalizeString(galleryFilter.filterSearch).trim() != "";
-const emitter = inject<Emitter<Events>>("emitter");
 const heartbeat = storeHeartbeat();
 const configStore = storeConfig();
-
-socket.on(
-  "scan:scanning_platform",
-  ({ name, slug, id }: { name: string; slug: string; id: number }) => {
-    scanningStore.set(true);
-    scanningPlatforms.value = scanningPlatforms.value.filter(
-      (platform) => platform.name !== name
-    );
-    scanningPlatforms.value.push({ name, slug, id, roms: [] });
-  }
-);
-
-socket.on("scan:scanning_rom", (rom: SimpleRom) => {
-  scanningStore.set(true);
-  if (romsStore.currentPlatform?.id === rom.platform_id) {
-    romsStore.add([rom]);
-    romsStore.setFiltered(
-      isFiltered ? romsStore.filteredRoms : romsStore.allRoms,
-      galleryFilter
-    );
-  }
-
-  let scannedPlatform = scanningPlatforms.value.find(
-    (p) => p.slug === rom.platform_slug
-  );
-
-  // Add the platform if the socket dropped and it's missing
-  if (!scannedPlatform) {
-    scanningPlatforms.value.push({
-      name: rom.platform_name,
-      slug: rom.platform_slug,
-      id: rom.platform_id,
-      roms: [],
-    });
-    scannedPlatform = scanningPlatforms.value[0];
-  }
-
-  scannedPlatform?.roms.push(rom);
-});
-
-socket.on("scan:done", () => {
-  scanningStore.set(false);
-  socket.disconnect();
-
-  emitter?.emit("refreshDrawer", null);
-  emitter?.emit("snackbarShow", {
-    msg: "Scan completed successfully!",
-    icon: "mdi-check-bold",
-    color: "green",
-    timeout: 4000,
-  });
-});
-
-socket.on("scan:done_ko", (msg) => {
-  scanningStore.set(false);
-
-  emitter?.emit("snackbarShow", {
-    msg: `Scan failed: ${msg}`,
-    icon: "mdi-close-circle",
-    color: "red",
-  });
-  socket.disconnect();
-});
 
 onMounted(async () => {
   await api.get("/heartbeat").then(({ data: data }) => {
@@ -97,13 +21,6 @@ onMounted(async () => {
   await api.get("/config").then(({ data: data }) => {
     configStore.set(data);
   });
-});
-
-onBeforeUnmount(() => {
-  socket.off("scan:scanning_platform");
-  socket.off("scan:scanning_rom");
-  socket.off("scan:done");
-  socket.off("scan:done_ko");
 });
 </script>
