+++ frontend/src/components/Administration/Users/Dialog/CreateUser.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
+import RDialog from "@/components/common/RDialog.vue";
 import userApi from "@/services/api/user";
 import storeUsers from "@/stores/users";
 import type { Events } from "@/types/emitter";
 import type { Emitter } from "mitt";
 import { inject, ref } from "vue";
 import { useDisplay } from "vuetify";
-import RDialog from "@/components/common/RDialog.vue";
 
 // Props
 const user = ref({
