+++ frontend/src/plugins/router.ts
@@ -74,6 +74,7 @@ const router = createRouter({
 router.afterEach(() => {
   // Scroll to top to avoid annoying behaviour in mobile
   window.scrollTo({ top: 0, left: 0 });
+  // TODO: check permission for views. Ex: view user can access to /scan view
 });
 
 export default router;
